+++ apps/gizual-app/src/pages/main/main.module.scss.d.ts
@@ -3,6 +3,6 @@ declare const classNames: {
   readonly Page: "Page";
   readonly canvas: "canvas";
   readonly TitleBarContainer: "TitleBarContainer";
-  readonly AnalysePage: "AnalysePage";
+  readonly AnalyzePage: "AnalyzePage";
 };
 export = classNames;