+++ apps/gizual-app/src/pages/main/main.tsx
@@ -33,7 +33,7 @@ export const MainPage = observer(({ vm: externalVm }: MainPageProps) => {
 
       <div className={style.Body}>
         {mainController.selectedPanel === "explore" && <ExplorePage vm={vm} />}
-        {mainController.selectedPanel === "analyse" && <AnalysePage vm={vm} />}
+        {mainController.selectedPanel === "analyze" && <AnalyzePage vm={vm} />}
         {mainController.selectedPanel === "settings" && <SettingsPage />}
       </div>
     </div>
@@ -52,7 +52,7 @@ const ExplorePage = observer(({ vm }: MainPageProps) => {
   );
 });
 
-const AnalysePage = observer(({ vm }: MainPageProps) => {
+const AnalyzePage = observer(({ vm }: MainPageProps) => {
   const mainController = useMainController();
   if (!vm) return <div />;
   const [width, height] = useWindowSize();
@@ -72,7 +72,7 @@ const AnalysePage = observer(({ vm }: MainPageProps) => {
   const languageData = parseLanguages(mainController.fileTreeRoot);
 
   return (
-    <div ref={ref} className={style.AnalysePage}>
+    <div ref={ref} className={style.AnalyzePage}>
       <ReactGridLayout layout={layout} width={canvasWidth} cols={6} rowHeight={canvasWidth / 5}>
         <div key={"a"}>
           <Container title={"Language Distribution"}>