+++ apps/gizual-app/src/primitives/file-tree/file-tree.vm.tsx
@@ -32,22 +32,22 @@ export class FileTreeViewModel {
     return this._mainController.selectedFiles;
   }
 
-  get selectedFavouriteFiles(): string[] {
+  get selectedFavoriteFiles(): string[] {
     const files: string[] = [];
     for (const file of this.selectedFiles) {
-      if (this._mainController.favouriteFiles.has(file)) files.push(file);
+      if (this._mainController.favoriteFiles.has(file)) files.push(file);
     }
     return files;
   }
 
-  toggleFavourite(name: string) {
-    this._mainController.toggleFavourite(name);
+  toggleFavorite(name: string) {
+    this._mainController.toggleFavorite(name);
   }
 
-  get favouriteTreeData(): FileTreeDataNode[] {
+  get favoriteTreeData(): FileTreeDataNode[] {
     const tree: FileTreeDataNode[] = [];
 
-    for (const [path, info] of this._mainController.favouriteFiles) {
+    for (const [path, info] of this._mainController.favoriteFiles) {
       if (!info) continue;
       tree.push({
         key: path,
@@ -163,12 +163,12 @@ export class FileTreeViewModel {
     this._mainController.vmController?.canvasViewModel?.zoomToFile(path);
   }
 
-  setFavourite(node: FileTreeDataNode) {
-    this._mainController.toggleFavourite(node.path, node);
+  setFavorite(node: FileTreeDataNode) {
+    this._mainController.toggleFavorite(node.path, node);
   }
 
-  isFavourite(node: FileTreeDataNode) {
-    return this._mainController.favouriteFiles.has(node.path);
+  isFavorite(node: FileTreeDataNode) {
+    return this._mainController.favoriteFiles.has(node.path);
   }
 
   isFileSelected(node: FileTreeDataNode) {