+++ apps/gizual-app/src/primitives/icon-button/icon-button.module.scss.d.ts
@@ -2,6 +2,6 @@ declare const classNames: {
   readonly IconButton: "IconButton";
   readonly Wide: "Wide";
   readonly BorderRight: "BorderRight";
-  readonly Coloured: "Coloured";
+  readonly Colored: "Colored";
 };
 export = classNames;