+++ apps/gizual-app/src/types/visualization.ts
@@ -1,5 +1,5 @@
-export type VisualisationConfig = {
-  colours: {
+export type VisualizationConfig = {
+  colors: {
     oldest: string;
     newest: string;
     notLoaded: string;