+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 
 ### Changes
 
-- **app:file-tree:** improve partial select state behaviour
+- **app:file-tree:** improve partial select state behavior
 
 ---
 
@@ -54,13 +54,13 @@
 ### Features
 
 - **app:timeline:** add option to collapse timeline into search-bar
-- **app:canvas:** append colour selection and author panel to canvas stage
+- **app:canvas:** append color selection and author panel to canvas stage
 
 ### Changes
 
 - **app:general:** remove repo-panel and settings-panel
 - **app:search-bar:** modularize search-bar assistants
-- **app:search-bar:** improve tag update and synchronization behaviour
+- **app:search-bar:** improve tag update and synchronization behavior
 
 ### Bug Fixes